requirements.txt
pandas
SQLAlchemy
PyMySQL
python-dotenv
cryptography